requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure.functions
azure.identity
azure.mgmt.resourcegraph
azure.mgmt.resource
